# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "CRC PDF Generator",
    "contact": {
      "name": "RedHatInsights",
      "email": "support@redhat.com"
    },
    "description": "CRC PDF Generator API Documentation",
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "/api/crc-pdf-generator"
    }
  ],
...
  "paths": {
...
    "/v2/download/{ID}": {
      "get": {
        "tags": [
          "pdf"
        ],
        "operationId": "downloadPdf",
        "parameters": [
          {
            "name": "ID",
            "in": "path",
            "description": "ID of Pdf Collection to be downloaded",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "description": "Fetch the rendered PDF by the ID",
        "responses": {
          "200": {
            "$ref": "#/components/responses/RenderedPdf"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "403": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    }
  },
  "components": {
    "responses": {
      "Unauthorized": {
        "description": "Insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error403"
            }
          }
        }
      },
      "BadRequest": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error400"
            }
          }
        }
      },
      "NotFound": {
        "description": "404 Not Found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error404"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error500"
            }
          }
        }
      },
      "RenderedPdf": {
        "description": "Returns a PDF generated based on the service and template in the query string",
        "content": {
          "application/pdf": {
            "schema": {
              "$ref": "#/components/schemas/RenderedPdf"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error403": {
        "type": "object",
        "description": "Error schema for 403 HTTP code",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "Error400": {
        "type": "object",
        "description": "Error schema for 400 HTTP code",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "status": {
                "type": "string"
              },
              "statusText": {
                "type": "string"
              },
              "description": {
                "type": "string"
              }
            }
          }
        }
      },
      "Error404": {
        "type": "object",
        "description": "Error schema for 404 HTTP code",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "status": {
                "type": "string"
              },
              "statusText": {
                "type": "string"
              },
              "description": {
                "type": "string"
              }
            }
          }
        }
      },
      "Error500": {
        "type": "object",
        "description": "Error schema for 500 HTTP code",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "status": {
                "type": "string"
              },
              "statusText": {
                "type": "string"
              },
              "description": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "RenderedPdf": {
        "description": "Pdf stream returned as media type pdf",
        "type": "string",
        "format": "binary"
      }
    }
  }
}